Added method to store text
