--- signposts-again/View/NewSignView.swift
@@ -1,20 +1,0 @@
-//
-//  NewSignView.swift
-//  signposts-again
-//
-//  Created by Louis Kirkham on 10/11/2020.
-//
-
-import SwiftUI
-
-struct NewSignView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct NewSignView_Previews: PreviewProvider {
-    static var previews: some View {
-        NewSignView()
-    }
-}
