+++ signposts/Assets.xcassets/Contents.json
