+++ signposts/Assets.xcassets/signpostsyellow.imageset/Contents.json
