+++ signposts/Assets.xcassets/signpostsyellow.imageset/signpostsyellow.png
