+++ signpostsUITests/Info.plist
